refactor(interop): extract shared nodeRunner for Compose Tier 2
Pull worker lifecycle and command handlers behind a transport seam so
Docker Compose can reuse the same node entrypoint without forking interop logic.
Also close the adaptive scheduling chapter docs and add the Tier 2 plan.
