+++ BACKLOG.md
@@ -114,7 +114,7 @@ The core architecture has successfully transitioned to a robust, Dependency-Inje
 > **Implementation plan:** [tasks/tier1-adaptive-interop.md](tasks/tier1-adaptive-interop.md) — groomed checklist, CI strategy, phased PRs.
 
 * **Severity:** Medium (confidence / regression prevention for adaptive scheduler)
-* **Status:** PR merge gate complete ([PR #37](https://github.com/Ahmed-Abdul-Rahman/de-chat/pull/37) CI green) — nightly soak pending after merge
+* **Status:** **Completed** (2026-07-16) — [PR #37](https://github.com/Ahmed-Abdul-Rahman/de-chat/pull/37) + consecutive nightly greens; Bandit [PR #39](https://github.com/Ahmed-Abdul-Rahman/de-chat/pull/39) merged
 * **Depends on:** Task 2.1 adaptive scheduling (PR #36) — **done**
 * **Goal:** Increase statistical confidence in adaptive anti-entropy behaviour **without** new infrastructure. Build on the existing single-host worker-thread interop that already runs in CI.
 
@@ -182,8 +182,8 @@ The core architecture has successfully transitioned to a robust, Dependency-Inje
 
 - [x] PR CI (`test:int:data-sync`) passes with `ADAPTIVE_INTEROP_STRICT=true` ([PR #37](https://github.com/Ahmed-Abdul-Rahman/de-chat/pull/37))
 - [x] Interop report prints `idleSkips`, `usefulSyncs`, `floorSyncForces` for late joiner
-- [ ] A/B scenario: both modes converge; heuristic shows `idleSkips > 0` on dormant producers (nightly, post-merge)
-- [ ] Nightly 50-node adaptive run completes without worker-thread leak (post-merge)
+- [x] A/B scenario: both modes converge; heuristic shows `idleSkips > 0` on dormant producers (nightly)
+- [x] Nightly 50-node adaptive run completes without worker-thread leak (post-merge soak)
 - [x] No change to production `adaptive.enabled` default (`false`)
 
 #### Out of scope (Tier 1)
@@ -196,9 +196,11 @@ The core architecture has successfully transitioned to a robust, Dependency-Inje
 
 ### Task 5.2: Tier 2 — Docker Compose Interop (Real Network Isolation at Scale)
 
+> **Implementation plan:** [tasks/tier2-compose-interop.md](tasks/tier2-compose-interop.md) — groomed checklist, phased PRs, approval gate.
+
 * **Severity:** Medium–High (realistic P2P conditions; catches bugs invisible on localhost worker threads)
-* **Status:** Backlog
-* **Depends on:** Task 5.1 (strict polling + enriched metrics) — recommended first
+* **Status:** **Active** — plan ready for approval (2026-07-16); Tier 1 + Bandit closed
+* **Depends on:** Task 5.1 — **done**
 * **Goal:** Run the same DeChat convergence scenarios in **isolated containers** with configurable network conditions (latency, jitter, bandwidth, partitions), following the pattern libp2p adopted after leaving Testground.
 
 #### Why Tier 2 (not Testground)
